requirements.txt
discord.py
giphy_client
bs4